Make properties() give a reference to the internal hash
This change stops properties() in Node and Relationship from making a new defensive copy every time it's called. This is 10 % to 60 % faster in my tests; however, retrieving a property is very fast anyway, so this speed difference may not matter much.

More importantly, though, the resulting code is also clearer: From the name of the get() method, it isn't obvious that it's getting a property, and the Record class actually has a get() method of its own that does something entirely different. In the past, this often lead to code like "->get->get", which looks like a typo and is somewhat confusing.

Finally, Neo4j::Bolt has always done this, and this change makes values received over HTTP behave the same way.