Change docs for Bolt availability from Neo4j 3.0 to 3.5
This is not a functional change. Neo4j::Driver 1.x continues to target all versions of Neo4j 3.

But versions up until Neo4j 3.4 require Bolt v1, which is hardly tested and may not be very reliable. To the extent that perlbolt & friends provide Bolt v1 support, it should continue to work. This change merely intends to remind users of really old Neo4j versions that it may be time to upgrade, at least to Neo4j 3.5.